chore(kobo): merge with `kobo_sync_utils`
